Move views into folders and remove old routes
