+++ web/src/views/HomeView/HomeView.vue
@@ -1,0 +1,11 @@
+<template>
+  <v-container>
+    TODO
+  </v-container>
+</template>
+
+<script>
+export default {
+  name: 'HomeView',
+};
+</script>
